final nits and moving some files
