+++ common/currency_list.go
@@ -1,8 +1,12 @@
-package sleet
+package common
+
+import (
+	"github.com/BoltApp/sleet"
+)
 
 // CURRENCIES maps the precision to the currency symbol used for lookups for some PsP providers that rely on these values for amount calculation
 // One example is Braintree which uses its own amount structure requiring the precision of a currency
-var CURRENCIES = map[Code]Currency{
+var CURRENCIES = map[Code]sleet.Currency{
 	AED: {Precision: 2, Symbol: "AED"},
 	AFN: {Precision: 2, Symbol: "؋"},
 	ALL: {Precision: 2, Symbol: "Lek"},
