+++ common/units_of_measurement.go
@@ -1,4 +1,4 @@
-package sleet
+package common
 
 import "strings"
